fix: improving logs

+++ server/src/controllers/list.controller.ts
@@ -1,4 +1,3 @@
-import chalk from "chalk";
 import { Request, Response } from "express";
 import { Types } from "mongoose";
 import {
@@ -7,7 +6,7 @@ import {
   joinToListService,
   leaveListService,
 } from "../services/list.service";
-import { deleteListDal, updateListNameDal } from "../dal/list.dal";
+import { deleteListDal } from "../dal/list.dal";
 import { updateListNameService } from "../services/product.service";
 import { CustomError, handleError } from "../utils/handleError";
 import logger from "../helpers/logger";
@@ -20,18 +19,8 @@ export const createList = async (req: Request, res: Response) => {
     if (!Types.ObjectId.isValid(userId))
       throw new CustomError("Invalid userId", 400, "LIST-ERROR");
 
-    // בדיקה אם המשתמש כבר יצר רשימה עם אותו שם
-    const existing = await checkingUserInListSerice(listName, userId);
-    if (existing)
-      throw new CustomError(
-        "List name already exists for this user",
-        400,
-        "LIST-ERROR"
-      );
-
     // יצירת הרשימה והוספת המשתמש שיוצר אותה
     const newList = await createListService(listName, userId);
-
     res.json(newList);
   } catch (error) {
     return handleError(res, error);
@@ -45,6 +34,17 @@ export const joinToList = async (req: Request, res: Response) => {
 
     if (!Types.ObjectId.isValid(listId) || !Types.ObjectId.isValid(userId))
       throw new CustomError("Invalid listId or userId", 400, "LIST-ERROR");
+
+    // בדיקה אם המשמש כבר קיים ברשימה
+    const existing = await checkingUserInListSerice(listId, userId);
+        console.log(1, existing);
+
+    if (existing)
+      throw new CustomError(
+        "The user is already on the list.",
+        400,
+        "LIST-ERROR"
+      );
 
     const updatedList = await joinToListService(listId, userId);
     res.json(updatedList);
@@ -60,12 +60,23 @@ export const leaveList = async (req: Request, res: Response) => {
 
     if (!Types.ObjectId.isValid(listId) || !Types.ObjectId.isValid(userId))
       throw new CustomError("Invalid listId or userId", 400, "LIST-ERROR");
+    const existing = await checkingUserInListSerice(listId, userId);
 
+    if (!existing)
+      throw new CustomError(
+        "The user does not exist in the list.",
+        400,
+        "LIST-ERROR"
+      );
     // הסרת המשתמש מהרשימה
     const updatedList = await leaveListService(listId, userId);
-    // אם אין יותר משתמשים ברשימה → מחיקה אוטומטית
+    // אם אין יותר משתמשים ברשימה  מחיקה אוטומטית
     if (updatedList?.userIds.length === 0) {
-      await deleteListDal(listId);
+      const deleteList = await deleteListDal(listId);
+
+      if (deleteList) {
+        logger("HTTP", `List with id ${listId} deleted because it was empty after user left.`,'greenBright', "blueBright");
+      }
 
       // return res.json({
       //   message: "User left and list was deleted because it's empty",
